requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
asgiref==3.5.0
    # via
    #   -r requirements/base.txt
    #   django
attrs==21.4.0
    # via pytest
bleach[css]==5.0.0
    # via -r requirements/base.txt
coverage[toml]==6.3.2
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   django-classy-tags
    #   django-js-asset
    #   django-sekizai
django-classy-tags==3.0.1
    # via
    #   -r requirements/base.txt
    #   django-sekizai
django-js-asset==2.0.0
    # via
    #   -r requirements/base.txt
    #   django-mptt
django-mptt==0.13.4
    # via -r requirements/base.txt
django-sekizai==3.0.1
    # via -r requirements/base.txt
importlib-metadata==4.11.3
    # via
    #   -r requirements/base.txt
    #   markdown
iniconfig==1.1.1
    # via pytest
markdown==3.3.6
    # via -r requirements/base.txt
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pyparsing==3.0.8
    # via packaging
pytest==7.1.1
    # via
    #   -r requirements/test.in
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
pytz==2022.1
    # via
    #   -r requirements/base.txt
    #   django
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   bleach
sorl-thumbnail==12.8.0
    # via -r requirements/base.txt
sqlparse==0.4.2
    # via
    #   -r requirements/base.txt
    #   django
tinycss2==1.1.1
    # via
    #   -r requirements/base.txt
    #   bleach
tomli==2.0.1
    # via
    #   coverage
    #   pytest
webencodings==0.5.1
    # via
    #   -r requirements/base.txt
    #   bleach
    #   tinycss2
zipp==3.8.0
    # via
    #   -r requirements/base.txt
    #   importlib-metadata
